feat(rhythm): support fractional counts for long-period habits
Remove the .int() constraint from the rhythm count schema so habits
like passport renewal can use { period: "annually", count: 0.1 }
instead of requiring new period types.

+++ mcp-server/vault.ts
@@ -63,7 +63,7 @@ export type RhythmPeriod = z.infer<typeof RhythmPeriodSchema>;
 
 export const RhythmSchema = z.object({
   period: RhythmPeriodSchema,
-  count: z.number().int().positive(),
+  count: z.number().positive(),
 });
 export type Rhythm = z.infer<typeof RhythmSchema>;
